Add files via upload
UNIPLIX is the UI of a website that lets space lovers explore the mirage of the universe made by humans. It is an early-stage collection of informative movies and documentaries that help in analyzing the nature of the universe. This includes easy navigation, visually appealing images and videos, personalized recommendations, user-friendly search functionality, and intuitive content categorization. The celestial theme should be integrated throughout the UI, with a color scheme and visual design that is consistent and visually appealing.

Challenges we ran into
Designing for a platform as large and complex as Uniplix presents a number of challenges. Here are some of the key challenges faced by the Uniplix design team:

Software Limitations: Figma is no doubt a good UI designing app, but it restricts certain interactive and responsive features. It is time-consuming as certain interactive elements are not available. this caused a major impact on the interactive quality of our design. We could have made the program much better and more interactive with data storing capacities only if multiple software choices were given, but we accepted that as a challenge and did our best.

Narration feature: We tried to append our system by adding a feature for the visually impaired people by adding voiceovers, but then for the voiceovers to proceed we needed some kind of interaction with the user, which was not basically possible in this case.
So we came up with the idea of connecting our system with braille keyboards so that even visually impaired people can access our website by reacting through the braille keyboard. We will work on this in the future as time was a constraint for now.

Data-driven design: Uniplix is highly data-driven, and this requires a deep understanding of data analysis tools and techniques. But as earlier said in our case we were not provided with data handling components in the software so for now we made it completely UI.
Overall, designing for uniplix requires a skilled and flexible design team that is able to navigate the complexities of a large and constantly evolving platform. By focusing on scalability, consistency, accessibility, localization, and data-driven design, the UNIPLIX design team will be able to create a highly engaging and user-friendly platform that meets the needs of millions of users worldwide in the future.
DO SCROLL HORIZONTALLY
https://www.figma.com/proto/4OXQAIutSbbEvvDd7A0PHp/UNIPLIX?node-id=1-2&scaling=scale-down&page-id=0%3A1&starting-point-node-id=1%3A2&show-proto-sidebar=1